+++ src/parser/state.rs
@@ -345,7 +345,7 @@ fn state_ref(input: Input<'_>) -> IResult<Input<'_>, Node<RefDecl>> {
                 is_derived: prefix.is_derived,
                 usage_prefix: prefix.usage_prefix,
                 is_constant: prefix.is_constant,
-                direction: None,
+                direction: prefix.direction,
                 kind_keyword: None,
                 name: name_str,
                 typing,